+++ Sources/TestHelper/RouterDelegateSpy.swift
@@ -2,6 +2,7 @@ import Foundation
 import MapboxCoreNavigation
 import CoreLocation
 import MapboxDirections
+import Turf
 
 public final class RouterDelegateSpy: RouterDelegate {
     public var onDidRefresh: ((RouteProgress) -> Void)?
@@ -19,6 +20,8 @@ public final class RouterDelegateSpy: RouterDelegate {
     public var onDidArriveAt: ((Waypoint) -> Bool)?
     public var onShouldPreventReroutesWhenArrivingAt: ((Waypoint) -> Bool)?
     public var onRouterShouldDisableBatteryMonitoring: (() -> Bool)?
+    public var onManeuverBufferWhenRerouting: (() -> LocationDistance?)?
+    public var onReroutingRequest: ((RouteOptions) -> ReroutingRequestBehavior)?
 
     public init() {}
 
@@ -35,7 +38,15 @@ public final class RouterDelegateSpy: RouterDelegate {
                        willRerouteFrom location: CLLocation) {
         onWillRerouteFrom?(location)
     }
+    
+    public func router(_ router: Router, initialManeuverBufferWhenReroutingFrom location: CLLocation) -> LocationDistance? {
+        return onManeuverBufferWhenRerouting?() ?? RouteController.DefaultBehavior.reroutingManeuverRadius
+    }
 
+    public func router(_ router: Router, requestBehaviorForReroutingWith options: RouteOptions) -> ReroutingRequestBehavior {
+        return onReroutingRequest?(options) ?? .default
+    }
+    
     public func router(_ router: Router,
                        shouldDiscard location: CLLocation) -> Bool {
         return onShouldDiscard?(location) ?? RouteController.DefaultBehavior.shouldDiscardLocation